lib/diary_entry.py
import math
class DiaryEntry:
    def __init__(self, title, contents):
        # Parameters:
        #   title: string
        #   contents: string
        self.title = title
        self.contents = contents
        self.reading_index = 0

    def format(self):
        # Returns:
        #   A formatted diary entry, for example:
        #   "My Title: These are the contents"
        return f"{self.title}: {self.contents}"

    def count_words(self):
        # Returns:
        #   int: the number of words in the diary entry
        count = len(self.title.split())
        count += len(self.contents.split())
        return count

    def reading_time(self, wpm):
        # Parameters:
        #   wpm: an integer representing the number of words the user can read 
        #        per minute
        # Returns:
        #   int: an estimate of the reading time in minutes for the contents at
        #        the given wpm.
        if wpm <= 0:
            raise Exception("You must read more than 0 words per minute")
        words = self.count_words()
        return math.ceil(words / wpm)

    def reading_chunk(self, wpm, minutes):
        # Parameters
        #   wpm: an integer representing the number of words the user can read
        #        per minute
        #   minutes: an integer representing the number of minutes the user has
        #            to read
        # Returns:
        #   string: a chunk of the contents that the user could read in the
        #           given number of minutes
        #
        # If called again, `reading_chunk` should return the next chunk,
        # skipping what has already been read, until the contents is fully read.
        # The next call after that should restart from the beginning.
        words_list = self.contents.split()
        number_words_to_be_read = wpm * minutes
        text = " ".join(words_list[self.reading_index:self.reading_index + number_words_to_be_read])
        self.reading_index += number_words_to_be_read
        if self.reading_index >= len(words_list):
            self.reading_index = 0
        return text
    